fix: better file structure
